send more user data if requested by an admin
